Move HighlightCardsHeading inline
